# [aggregator/ckey] implement a faster hash to reduce CPU and RAM usage 