--- a/pkg/aggregator/sketch_map_test.go
+++ b/pkg/aggregator/sketch_map_test.go
@@ -28,8 +28,10 @@ func TestInsert(t *testing.T) {
 		SampleRate: 1,
 	}
 
-	sketchMap.insert(1, generateContextKey(&mSample1), 1, 1)
+	ckey, _ := generateContextKey(&mSample1)
+	sketchMap.insert(1, ckey, 1, 1)
 	assert.Equal(t, 1, sketchMap.Len())
-	sketchMap.insert(2, generateContextKey(&mSample1), 2, 1)
+	ckey, _ = generateContextKey(&mSample1)
+	sketchMap.insert(2, ckey, 2, 1)
 	assert.Equal(t, 2, sketchMap.Len())
 }