--- a/pkg/dogstatsd/server_test.go
+++ b/pkg/dogstatsd/server_test.go
@@ -608,11 +608,11 @@ func TestDebugStats(t *testing.T) {
 	sample3 := metrics.MetricSample{Name: "some.metric3", Tags: make([]string, 0)}
 	sample4 := metrics.MetricSample{Name: "some.metric4", Tags: []string{"b", "c"}}
 	sample5 := metrics.MetricSample{Name: "some.metric4", Tags: []string{"c", "b"}}
-	hash1 := keygen.Generate(sample1.Name, "", sample1.Tags)
-	hash2 := keygen.Generate(sample2.Name, "", sample2.Tags)
-	hash3 := keygen.Generate(sample3.Name, "", sample3.Tags)
-	hash4 := keygen.Generate(sample4.Name, "", sample4.Tags)
-	hash5 := keygen.Generate(sample5.Name, "", sample5.Tags)
+	hash1, _ := keygen.Generate(sample1.Name, "", sample1.Tags)
+	hash2, _ := keygen.Generate(sample2.Name, "", sample2.Tags)
+	hash3, _ := keygen.Generate(sample3.Name, "", sample3.Tags)
+	hash4, _ := keygen.Generate(sample4.Name, "", sample4.Tags)
+	hash5, _ := keygen.Generate(sample5.Name, "", sample5.Tags)
 
 	// test ingestion and ingestion time
 	s.storeMetricStats(sample1)