--- a/releasenotes/notes/faster-ckey-hash-8f30ad530cce348a.yaml
+++ b/releasenotes/notes/faster-ckey-hash-8f30ad530cce348a.yaml
@@ -1,0 +1,12 @@
+# Each section from every release note are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+enhancements:
+  - |
+    Implemented a faster context key/hash generator: it should result in a reduced CPU
+    usage and a smaller RAM footprint for DogStatsD servers receiving a lot of metrics.