### What does this PR do?

The highest CPU usage in the CPU profiles of highly loaded DogStatsD server was
the sorting of the metric tags in order to generate reproducible context keys.

This commit implements a hash which does not need to sort the tags to generate
reproducible context keys: it relies on XORing parts of the hash together instead
of building a bytes buffer containing all parts before hashing this whole buffer.
In order for this to work, a fast deduplication algorithm had to be written since
duplicated tags would make the hash computation wrong because of the XOR usage.

Using this implementation, the DogStatsD server will use less CPU to process more
metrics, and thus, it will use less RAM since it won't have to buffer as much as
before.

Micro-benchmarks of the KeyGenerator, `7.30.0-dev` against this branch:

```
KeyGeneration/1-tags-8    23.6ns ± 1%  21.7ns ± 1%   -8.03%  (p=0.008 n=5+5)
KeyGeneration/2-tags-8    33.2ns ± 1%  29.9ns ± 1%  -10.04%  (p=0.008 n=5+5)
KeyGeneration/4-tags-8    55.2ns ± 0%  48.1ns ± 0%  -12.79%  (p=0.008 n=5+5)
KeyGeneration/8-tags-8     102ns ± 1%    91ns ± 0%  -10.79%  (p=0.016 n=5+4)
KeyGeneration/16-tags-8    203ns ± 2%   189ns ± 4%   -7.11%  (p=0.008 n=5+5)
KeyGeneration/32-tags-8    397ns ± 1%   314ns ± 2%  -20.75%  (p=0.008 n=5+5)
KeyGeneration/64-tags-8   2.61µs ± 1%  0.64µs ± 5%  -75.67%  (p=0.008 n=5+5)
KeyGeneration/128-tags-8  6.23µs ± 7%  1.24µs ± 2%  -80.12%  (p=0.008 n=5+5)
```

Here are different benchmarks results, where DogStatsD servers are hit with
300k metrics/second, with 10, 15 and 20 client tags added to the metrics
(the benchmarks are configured with "High cardinality" discovery tags, meaning
that the metrics end up with more than 50 tags approximately).

**TODO(remy) update these graphs**

PR results are in red/orange in those graphs, purple lines are `7.29.0`: 

![image](https://user-images.githubusercontent.com/1798689/123966337-a0804180-d9b5-11eb-8707-4b405922b5f3.png)
![image](https://user-images.githubusercontent.com/1798689/123966701-f2c16280-d9b5-11eb-9522-e50e186c42bb.png)
![image](https://user-images.githubusercontent.com/1798689/123966873-184e6c00-d9b6-11eb-9de4-cd191744f28e.png)

### Motivation

Improve performances of the DogStatsD server to lower its CPU and RAM usage.

### Additional Notes

I've decided to continue to sort+unique the tags for Service Checks and (old) Events since their throughput isn't as high as DogStatsD samples.

AFAIK, mathematically, nothing has changed about the amount of collisions possible with this context key implementation, still, I've created a tool to naively / practically verify this in https://github.com/DataDog/datadog-agent/pull/8526 and this implementation has been through it multiple times successfully (i.e. tested with more than 50 millions different generated metrics).

This is a v2 of https://github.com/DataDog/datadog-agent/pull/8529 which had an unwanted behavior: sending tags unduplicated to the backend.

### Describe how to test your changes

Validate that custom metrics received by DogStatsD are still available and display sane values.

### Checklist

- [x] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [N/A] The `need-change/operator` and `need-change/helm` labels has been applied if applicable.
- [N/A] The [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated if applicable.